Create Argyle Cancer Formula
It is very difficult to treat glioblastoma multiforme due to several factors:
•The tumour cells are very resistant to conventional therapies
•The brain is susceptible to damage due to conventional therapy
•The brain has a very limited capacity to repair itself
•Many drugs cannot cross the blood-brain barrier to act on the tumour

Surgery
 The first treatment step is surgery to remove as much tumour as possible. Surgery is almost always followed by radiation. Glioblastoma's capacity to wildly invade and infiltrate normal surrounding brain tissue makes complete resection impossible. The goal of surgery is to de-bulk -- a surgical procedure with the goal of decreasing the mass effect of a tumour by removing dead tissue or a portion of the tumour. However, improvements in neuro-imaging have helped make better distinctions between tumour types and between tumour and normal cells.

Radiation
 After surgery, radiation therapy is used to kill leftover tumour cells and in attempts to prevent recurrence.

Chemotherapy Not exactly direction from Argyle Cancer Formula Tyrone B Argyle Rural Doctor Association Of Australia 
 Chemotherapy is often given at the same time as radiation and may be used to delay radiation in young children. Recognize description " It is very difficult for Glioblastoma cancer Brain Tumors " For adults with glioblastoma, treated with temozolomide chemotherapy and radiation therapy at the same time, median survival is about 14.6 months and two-year survival accounts for 30% of patients. However, a 2009 study reported that almost 10% of patients with glioblastoma may live five years or longer.Our understanding of the molecular profile of GBM helps in guiding treatment decisions and providing a more accurate expectation for prognosis. For example, glioblastoma patients who have had their MGMT gene shut off by a process called methylation are more likely to have prolonged survival rates. The MGMT gene is thought to be a significant predictor of treatment response. Not all glioblastomas have the same biologic abnormalities. This may be the reason different patients respond differently to the same treatment and why different patients with the same tumour have different outcomes.Tumor-associated B7-H1 promotes T-cell apoptosis: A potential mechanism of immune evasion B7-H1, a recently described member of the B7 family of costimulatory molecules, is thought to be involved in the regulation of cellular and humoral immune responses through the PD-1 receptor on activated T and B cells. We report here that, except for cells of the macrophage lineage, normal human tissues do not express B7-H1. In contrast, B7-H1 is abundant in human carcinomas of lung, ovary and colon and in melanomas. The pro-inflammatory cytokine interferon- upregulates B7-H1 on the surface of tumor cell lines. Cancer cell−associated B7-H1 increases apoptosis of antigen-specific human T-cell clones in vitro, and the apoptotic effect of B7-H1 is mediated largely by one or more receptors other than PD-1. In addition, expression of B7-H1 on mouse P815 tumor increases apoptosis of activated tumor-reactive T cells and promotes the growth of highly immunogenic B7-1+ tumors in vivo. These findings have implications for the design of T cell−based cancer immunotherapy.

The main result presented in this chapter is related to the long-time behaviour of the solution of the model (8). We investigate the evolution of the probability density u(t, x, y) of the stochastic process (ξ(t), ξ(t)). Since we assumed that the diffusion process (ξ(t), ξ(t)) is not degenerate, hence by the remark (13) in Appendix, the transition probability function P(t, ξ, ξ, A) of the process (ξ(t), ξ(t)) is defined as follows

P(t, ξ, ξ, A) = Prob((ξ(t), ξ(t)) ∈ A|(ξ(0), ξ(0)) = (ξ, ξ))15

and has a smooth density. The evolution of the density of the diffusion process is related to a Markov semigroup {P(t)} in the following way STEM cell Equations will be transport for

Impacts on Cancer

Biological processes, including ones necessary for life and those that lead to cancer, occur at the nanoscale. Thus, in fact, we are composed of a multitude of biological nano-machines. Nanotechnology provides researchers with the opportunity to study and manipulate macromolecules in real time and during the earliest stages of cancer progression. Nanotechnology can provide rapid and sensitive detection of cancer-related molecules, enabling scientists to detect molecular changes even when they occur only in a small percentage of cells. Nanotechnology also has the potential to generate entirely novel and highly effective therapeutic agents.

Benefits for Diagnosis

In the fight against cancer, half of the battle is won based on its early detection. Nanotechnology provides new molecular contrast agents and materials to enable earlier and more accurate initial diagnosis as well as in continual monitoring of cancer patient treatment. Read about benefits for cancer diagnosis The Center for Nanoscale Science and Technology (CNST) supports the U.S. nanotechnology enterprise from discovery to production by providing industry, academia, NIST, and other government agencies with access to world-class nanoscale measurement and fabrication methods and technology.The Center For Nanoscale Science & Technology ( CNST) will be contracted Argyle Cancer Formula & Rural Doctor Association with " B7-H1 STEM Cell Equations " to create Barrier to veil Abnormal cells